# openapi.yaml
openapi: 3.1.0
info:
  title: Traccar
  description: Traccar GPS tracking server API documentation. To use the API you need to have a server instance. For testing purposes you can use one of free [demo servers](https://www.traccar.org/demo-server/). For production use you can install your own server or get a [subscription service](https://www.traccar.org/product/tracking-server/).
  contact:
    name: Traccar Support
    url: https://www.traccar.org/
    email: support@traccar.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '6.9.0'
servers:
  - url: https://demo.traccar.org/api
    description: Demo Server 1
  - url: https://demo2.traccar.org/api
    description: Demo Server 2
  - url: https://demo3.traccar.org/api
    description: Demo Server 3
  - url: https://demo4.traccar.org/api
    description: Demo Server 4
  - url: https://server.traccar.org/api
    description: Subscription Server
  - url: http://{host}:{port}/api
    description: Other Server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '8082'
          - '80'
        default: '8082'
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /drivers:
    get:
      summary: Fetch a list of Drivers
      description: Without params, it returns a list of Drivers the user has access to
      tags:
        - Drivers
      parameters:
        - name: all
          in: query
          description: Can only be used by admins or managers to fetch all entities
          schema:
            type: boolean
        - name: userId
          in: query
          description: Standard users can use this only with their own _userId_
          schema:
            type: integer
        - name: deviceId
          in: query
          description: >-
            Standard users can use this only with _deviceId_s, they have access
            to
          schema:
            type: integer
        - name: groupId
          in: query
          description: >-
            Standard users can use this only with _groupId_s, they have access
            to
          schema:
            type: integer
        - name: refresh
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Driver'
    post:
      summary: Create a Driver
      tags:
        - Drivers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Driver'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
      x-codegen-request-body-name: body
...
components:
  schemas:
...
    Driver:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        uniqueId:
          type: string
        attributes:
          type: object
          properties: {}
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer